+++ packages/card/src/Card.js
@@ -7,11 +7,18 @@ const Card = props => {
   return <StyledCard {...props} />;
 };
 
+Card.defaultProps = {
+  width: "100%",
+  padding: "16px",
+  overflow: "hidden"
+};
+
 Card.propTypes = {
   floating: PropTypes.bool,
   type: PropTypes.oneOf(Object.keys(cardTypes)),
   children: PropTypes.node,
   width: PropTypes.string,
-  padding: PropTypes.string
+  padding: PropTypes.string,
+  overflow: PropTypes.string
 };
 export default Card;