+++ packages/hoc-input/src/formInput.story.js
@@ -85,6 +85,16 @@ storiesOf("HOC formInput", module)
     withInfo()(() => <EnhancedInput type="search" value="broccoli" />)
   )
   .add(
+    "small html search input with value",
+    withInfo()(() => (
+      <EnhancedInput
+        type="search"
+        fontSize={fontSizes.SMALL}
+        value="broccoli"
+      />
+    ))
+  )
+  .add(
     "html search input with custom clear icon",
     withInfo()(() => (
       <EnhancedInput type="search" value="broccoli" clearIcon="wg-close" />
@@ -111,13 +121,16 @@ storiesOf("HOC formInput", module)
     ))
   )
   .add(
-    "custom renderContent",
+    "html input with leftIcon",
+    withInfo()(() => <EnhancedInput leftIcon="wg-order" value="foo bar" />)
+  )
+  .add(
+    "small html input with leftIcon",
     withInfo()(() => (
       <EnhancedInput
-        type="text"
-        renderContent={content => (
-          <div style={{ border: "dotted 4px hotpink" }}>{content}</div>
-        )}
+        leftIcon="wg-order"
+        fontSize={fontSizes.SMALL}
+        value="foo bar"
       />
     ))
   );