+++ packages/input-select/src/styledComponents/ItemImage.js
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-import Image from "@crave/farmblocks-image";
-
-export default styled(Image)`
-  margin: -6px;
-  margin-right: 8px;
-`;